requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

setuptools<70.0.0
torch==2.6.0
torchvision==0.21.0
decord

# SAM3
git+https://github.com/facebookresearch/sam3.git@f6e51f59500a87c576c2df2323ce56b9fd7a12de

# MatAnyone2
cython
gitpython>=3.1
thinplate@git+https://github.com/cheind/py-thin-plate-spline
hickle>=5.0
tensorboard>=2.11
numpy>=1.21
Pillow>=9.5
opencv-python>=4.8
scipy>=1.7
pycocotools>=2.0.7
tqdm>=4.66.1
gradio>=3.34
gdown>=4.7.1
einops>=0.6
hydra-core>=1.3.2
PySide6>=6.2.0
charset-normalizer>=3.1.0
netifaces-plus>=0.12.5
easydict
av>=0.5.2
requests
pyqtdarktheme
imageio==2.25.0
imageio[ffmpeg]
huggingface_hub
safetensors
xlsxwriter
kornia